move config_ui_repository into script_project_repository
