+++ lib/project_types/script/layers/infrastructure/push_package_repository.rb
@@ -7,7 +7,7 @@ module Script
         include SmartProperties
         property! :ctx, accepts: ShopifyCli::Context
 
-        def create_push_package(script_project:, script_content:, compiled_type:, metadata:, config_ui:)
+        def create_push_package(script_project:, script_content:, compiled_type:, metadata:)
           build_file_path = file_path(script_project.id, script_project.script_name, compiled_type)
           write_to_path(build_file_path, script_content)
 
@@ -18,11 +18,11 @@ module Script
             script_content: script_content,
             compiled_type: compiled_type,
             metadata: metadata,
-            config_ui: config_ui,
+            config_ui: script_project.config_ui,
           )
         end
 
-        def get_push_package(script_project:, compiled_type:, metadata:, config_ui:)
+        def get_push_package(script_project:, compiled_type:, metadata:)
           build_file_path = file_path(script_project.id, script_project.script_name, compiled_type)
           raise Domain::PushPackageNotFoundError unless ctx.file_exist?(build_file_path)
 
@@ -35,7 +35,7 @@ module Script
             script_content: script_content,
             compiled_type: compiled_type,
             metadata: metadata,
-            config_ui: config_ui,
+            config_ui: script_project.config_ui,
           )
         end
