+++ test/project_types/script/layers/infrastructure/fake_push_package_repository.rb
@@ -12,8 +12,7 @@ module Script
           script_project:,
           script_content:,
           compiled_type:,
-          metadata:,
-          config_ui:
+          metadata:
         )
           id = id(script_project.script_name, compiled_type)
           @cache[id] = Domain::PushPackage.new(
@@ -23,12 +22,12 @@ module Script
             script_content: script_content,
             compiled_type: compiled_type,
             metadata: metadata,
-            config_ui: config_ui,
+            config_ui: script_project.config_ui,
           )
         end
 
-        def get_push_package(script_project:, compiled_type:, metadata:, config_ui:)
-          _ = metadata, config_ui
+        def get_push_package(script_project:, compiled_type:, metadata:)
+          _ = metadata
           id = id(script_project.script_name, compiled_type)
           if @cache.key?(id)
             @cache[id]
